==> RQ4/0-Modify statistics multiple times.ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd
import re
import matplotlib.pyplot as plt
import numpy as np



path1 = 'D:/CCFinder/Commit2So(A commit only corresponds to the latest so)/'
list1 = os.listdir(path1)
len(list1),list1[0]

# list1


data_path = 'D:/CCFinder/NewData/'


import collections
class_info_dict = collections.defaultdict(list)
for file in tqdm(list1):
    name = file.split('D-CCFinder-NewData-')[1][:-4].split('-')
    commit_file = data_path+'-'.join(name[:-2])+'/'+'/'.join(name[-2:])

    with open(commit_file, encoding='utf-8') as f:
        lines = f.readlines()
        class_file = lines[1].rstrip('\n')
        split_class_file = class_file.split('/')
        class_name = split_class_file[-4]+'/'+split_class_file[-1]
        class_info_dict[class_name].append(file)



new_df = pd.DataFrame({'class_name':[], 'commit':[]})
for k, v in tqdm(class_info_dict.items()):
    if v==[]:print('error')
    new_df.loc[len(new_df)] = [k,
                               v]


new_df.to_csv('./The commit corresponding to the reused class.txt',index=False)

new_df = pd.read_csv('./The commit corresponding to the reused class.txt')
new_df.info()
