==> RQ4/00-exact match multiple modification is not the same modification.ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd
import re
import matplotlib.pyplot as plt
import numpy as np

df = pd.read_csv('./The commit corresponding to the reused class.txt')
df.info()
df['commit'] = df['commit'].apply(lambda x: eval(x))
df['prj'] = df['class_name'].apply(lambda x: x.split('/')[0])
df.head(3)

cc = df['commit'][1].copy()
cc

path1 = 'D:/CCFinder/testAns/'
for idx in tqdm(range(len(df))):
    project = df['prj'][idx]

    fileName = []
    clonePairs = []


    with open (path1 + '\\' + project + ".txt") as file:

        while 1:
            line = file.readline()
            if not line:
                break

            if line == 'source_files {\n':
                while 1:
                    fileLine = file.readline()
                    if fileLine == '}\n':
                        break
                    listName = fileLine.split('\t')
                    fileName.append(listName[1])

            if line == 'clone_pairs {\n':
                while 1:
                    PairsLine = file.readline()
                    if PairsLine == '}\n':
                        break
                    clonePairs.append(PairsLine)

    def readCommit(fileInfo):
        sp1 = fileInfo.split('.')
        fileIndex = int(sp1[0]) - 1
        fileName[fileIndex] = fileName[fileIndex].replace('F:\\', 'D:\\')
        return fileName[fileIndex]
    import collections
    hashTable = collections.defaultdict(list)

    for i in range(len(clonePairs)//2):

        listPair = clonePairs[i].split('\t')


        # listPair[1] = listPair[1].replace('F:\\', 'D:\\')
        Commitfile1 = readCommit(listPair[1])
        Commitfile2 = readCommit(listPair[2])
        Commitfile1 = Commitfile1.replace('\\', '-').replace(':','')+'.txt'
        Commitfile2 = Commitfile2.replace('\\', '-').replace(':','')+'.txt'

        if hashTable. __contains__(Commitfile1):
            # hashTable[SOfile] = hashTable[SOfile] + 1
            hashTable[Commitfile1][0] = hashTable[Commitfile1][0] + 1
            hashTable[Commitfile1].append(Commitfile2)
        else:
            # hashTable[SOfile] = 1
            hashTable[Commitfile1].append(1)
            hashTable[Commitfile1].append(Commitfile2)



    for commit in df['commit'][idx]:
        clone_list = hashTable[commit][1:]

        while commit in clone_list:
            clone_list.remove(commit)
        for clone_commit in clone_list:
            if clone_commit in df['commit'][idx]:
                df['commit'][idx].remove(clone_commit)



hashTable[cc[0]]

idx

df['commit'][idx]

cc


df.to_csv('./Commit corresponding to the reused class (exact).txt',index=False)


df['len'] = df['commit'].apply(lambda x: len(x))

df[df['len']>=2]

x = ['D-CCFinder-NewData-ActiveAndroid-2-746.java.txt',
 'D-CCFinder-NewData-ActiveAndroid-2-746.java.txt',
 'D-CCFinder-NewData-ActiveAndroid-2-746.java.txt',
 'D-CCFinder-NewData-ActiveAndroid-2-746.java.txt',
 'D-CCFinder-NewData-ActiveAndroid-2-746.java.txt',
 'D-CCFinder-NewData-ActiveAndroid-3-893.java.txt',
 'D-CCFinder-NewData-ActiveAndroid-3-893.java.txt',
 'D-CCFinder-NewData-ActiveAndroid-3-893.java.txt',
 'D-CCFinder-NewData-ActiveAndroid-3-893.java.txt',
 'D-CCFinder-NewData-ActiveAndroid-3-893.java.txt']



x.remove('D-CCFinder-NewData-ActiveAndroid-2-746.java.txt')

x
