==> RQ4/1-Modify statistics multiple times.ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd
import re
import matplotlib.pyplot as plt
import numpy as np


df1 = pd.read_csv('./Commit corresponding to the reused class (exact).txt')
df1.info()

df1['commit'] = df1['commit'].apply(lambda x: eval(x))
df1['len_commit'] = df1['commit'].apply(lambda x: len(x))
df1 = df1[df1['len_commit']>=2]
df1.index = list(range(len(df1)))
df1.info()

path = './A class that has been modified multiple times and has been reused multiple times/'
commit_path = 'D:/CCFinder/Commit2So(A commit only corresponds to the latest so)/'
for i in tqdm(range(len(df1))):
    file_name = df1['class_name'][i].replace('/', '-')
    with open(path+file_name+'.txt', 'w', encoding='utf-8') as f1:
        for commit in df1['commit'][i]:
            with open(commit_path+commit,'r',encoding='utf-8') as f2:
                content = f2.read()
                f1.write(content)


df1[df1['class_name']=='androidannotations/FromHtmlActivityTest.java']['commit'].to_list()
