==> RQ4/2-Continue analysis by modifying statistics multiple times.ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd
import re
import matplotlib.pyplot as plt
import numpy as np



so_post = '//CreationDate = 2019-01-15T05:49:12.343//Tags = <java><jtable><rowfilter>//Score = 1//Title = How to print specific Column values of the filtered Row//AnswerCount = 0//CommentCount = 0\n'






file_list = os.listdir('D:/CCFinder/CodeData/')
print('file nums: ', len(file_list))
file_list[:5]

flag = False
for file in tqdm(file_list):
    with open('D:/CCFinder/CodeData/'+file, 'r',encoding='utf-8') as f:
        lines = f.readlines()
        for line in lines:
            if line==so_post:
                flag = True
                print(file)
                break
    if flag==True:
        break



# Find classes that have been modified multiple times
# 'javassist-StringL.java'

# 'linuxdeploy-AppCompatPreferenceActivity.java'


# 'junit4-TextListener.java'

# 'maxwell-StringColumnDef.java'

# 'dubbo-LoggerAdapterTest.java'

path = 'D:/CCFinder/NewData/dubbo/'


for i in tqdm(os.listdir(path)):
    java_file_path = path + i + '/'
    java_file_list = os.listdir(java_file_path)
    for file in java_file_list:
        with open(java_file_path+file, 'r', encoding='utf-8') as f:
            content = f.read()
            if 'LoggerAdapterTest.java' in content:
                print(java_file_path+file)

java_file_list
