==> RQ4/2-Modify statistics multiple times.ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd
import re
import matplotlib.pyplot as plt
import numpy as np



path1 = './A class that has been modified multiple times and has been reused multiple times/'
list1 = os.listdir(path1)
len(list1),list1[0]

all_rereuse_post = []
sum = 0
exist_reuse_same_post_class_list = []
for file in tqdm(list1):
    with open(path1+file, 'r', encoding='utf-8') as f:
        lines = f.readlines()
        set_lines = list(set(lines))
        if len(set_lines) != len(lines):
            sum+=1
            exist_reuse_same_post_class_list.append(file)
        for line in set_lines:
            if lines.count(line)>=2:
                all_rereuse_post.append(line)
print('Modify and reuse the class of the same post multiple times：', sum)
print('proportion：', sum/len(list1))

len(all_rereuse_post)

len(exist_reuse_same_post_class_list)


exist_reuse_same_post_class_list

all_rereuse_post[-1]

len(all_rereuse_post)

with open('./Modify multiple posts that have been reused multiple times.txt', 'w', encoding='utf-8') as f:
    for i in all_rereuse_post:
        f.write(i)

x = [1,1,1,34,2]
x.count(1)

def extract_info(x):
    CreationDate = re.findall(r'CreationDate = (.*?)//',x)[0]

    Title = re.findall(r'Title = (.*?)//',x)[0]

    Tags = re.findall(r'Tags = (.*?)//',x)[0]
    Tags = Tags.split('><')
    Tags[0] = Tags[0].lstrip('<')
    Tags[-1] = Tags[-1].rstrip('>')
    Tags = ' '.join(Tags)

    Score = re.findall(r'Score = (.*?)//',x)[0]


    AnswerCount = re.findall(r'AnswerCount = (.*?)//',x)
    if AnswerCount==[] or AnswerCount==['']:
        AnswerCount = re.findall(r'AnswerCount = (.*?)\n',x)
        if AnswerCount == [] or AnswerCount==['']:
            AnswerCount = None
        else:
            AnswerCount = AnswerCount[0]
    else:
        AnswerCount = AnswerCount[0]


    CommentCount = re.findall(r'CommentCount = (.*?)//',x)
    if CommentCount==[] or CommentCount==['']:
        CommentCount = re.findall(r'CommentCount = (.*?)\n',x)
        if CommentCount==[] or CommentCount==['']:
            CommentCount = None
        else:
            CommentCount = CommentCount[0]
    else:
        CommentCount = CommentCount[0]

    FavoriteCount = re.findall(r'FavoriteCount = (.*?)//',x)
    if FavoriteCount==[] or FavoriteCount==['']:
        FavoriteCount = re.findall(r'FavoriteCount = (.*?)\n',x)
        if FavoriteCount==[] or FavoriteCount==['']:
            FavoriteCount = None
        else:
            FavoriteCount = FavoriteCount[0]
    else:
        FavoriteCount = FavoriteCount[0]



    return CreationDate, Title, Tags, Score, AnswerCount, CommentCount, FavoriteCount
new_df = pd.DataFrame({'CreationDate':[],'Title':[], 'Tags':[],
                          'Score':[],'AnswerCount':[], 'CommentCount':[],
                          'FavoriteCount':[]})

CreationDate_list = []
Title_list = []
Tags_list = []
Score_list = []
AnswerCount_list = []
CommentCount_list = []
FavoriteCount_list = []


with open('./Modify multiple posts that have been reused multiple times.txt', 'r', encoding='utf-8') as f:
    lines = f.readlines()
    # lines = list(set(lines))
    for l in lines:
        CreationDate, Title, Tags, Score, AnswerCount, CommentCount, FavoriteCount = extract_info(l)
        CreationDate_list.append(CreationDate)
        Title_list.append(Title)
        Tags_list.append(Tags)
        Score_list.append(Score)
        AnswerCount_list.append(AnswerCount)
        CommentCount_list.append(CommentCount)
        FavoriteCount_list.append(FavoriteCount)
new_df['CreationDate'] = CreationDate_list
new_df['Title'] = Title_list
new_df['Tags'] = Tags_list
new_df['Score'] = Score_list
new_df['AnswerCount'] = AnswerCount_list
new_df['CommentCount'] = CommentCount_list
new_df['FavoriteCount'] =FavoriteCount_list

new_df.to_csv('./Modify various indicators of posts that have been reused multiple times.txt')

new_df = pd.read_csv('./Modify various indicators of posts that have been reused multiple times.txt')
new_df.info()
